z64mm merge
merged z64mm into forked zeldaret repo

+++ include/libc/stdint.h
@@ -1,8 +1,10 @@
 #ifndef LIBC_STDINT_H
 #define LIBC_STDINT_H
 
-typedef signed long intptr_t;
-typedef unsigned long uintptr_t;
+#include "PR/ultratypes.h"
+
+typedef s32 intptr_t;
+typedef u32 uintptr_t;
 
 #define INT8_MIN    (-0x80)
 #define INT16_MIN   (-0x8000)
